adding comments and cleaning code up
