+++ app.py
@@ -1,26 +1,24 @@
-#%%
-import pandas as pd
+#%% Import Dependencies
 import streamlit as st
-import plotly.express as px
-import numpy as np
-import plotly.graph_objects as go
-import functions.acs_charts as fa
-import functions.acs_streamlit as ast
-import functions.acs_data_fetch as cf
-import functions.stocks_streamlit as ss
+st.set_page_config(layout="wide")
+import functions.housing_statistics.acs_charts as fa
+import functions.housing_statistics.acs_streamlit as ast
+import functions.housing_statistics.acs_data_fetch as cf
+import functions.stock_market.stocks_streamlit as ss
+import functions.tools as tl
 import us
-import time
 
-# Get a list of all US states
-us_states = [state.name for state in us.states.STATES]
+# Function to get a list of all US states
+st.cache_data(ttl='1d')
+def us_states_list():
+    us_states = [state.name for state in us.states.STATES]
+    return us_states
+us_states = us_states_list()
 
 # Get Most Recently Available ACS Year
 year = cf.get_most_recent_acs_year()
 year_max = year
 year_min = year - 10
-
-#%% Streamlit Setup
-st.set_page_config(layout="wide")
 
 # Read in Style.css
 with open('assets/style.css') as f:
@@ -29,11 +27,11 @@ with open('assets/style.css') as f:
 # Add Sidebar Filters
 dropdown_views = ['Housing Statistics', 'Stock Market']
 view_selection = st.sidebar.selectbox("#### View Selection", dropdown_views)
+st.sidebar.write('')
 
 
-
+# Create Housing Statistics Page
 if view_selection == 'Housing Statistics':
-
     # Display the page title
     st.markdown(
         f"""
@@ -43,51 +41,54 @@ if view_selection == 'Housing Statistics':
         """, unsafe_allow_html=True
     )
 
-    # Section 1 
-    st.markdown('#### Cumulative Change in Housing Costs and Incomes')
-    st.write('')
+    # Section 1: Cumulative Change in Housing Costs and Incomes
+    ## Section title
+    tl.write_around_markdown('#### Cumulative Change in Housing Costs and Incomes', 0, 1)
 
+    ## Create two columns to display filters side by side
     col1, col2 = st.columns([1, 2])
     with col1:
         state_selection = st.selectbox("**Select Geographic Level**", ['US National Average'] + us_states)
-
     with col2:
         year_range1 = st.slider('**Select Year Range**', min_value=year_min, max_value=year_max, value=(year_min, year_max), key='slider_key_1')
 
+    ## Display the Cumulative Change in Housing Costs and Incomes chart based on filter selections
     ast.yoy_cum_change_line_charts(state_selection, year_range1)
 
 
-    # Section 2
-    st.write('')
-    st.markdown('#### Share of Renters Housing Burdened Map')
-    st.write('')
+    # Section 2: Share of Renters Housing Burdened
+    ## Section title
+    tl.write_around_markdown('#### Share of Renters Housing Burdened Map', 1, 1)
 
-    # State vs. County Selection
+    ## State vs. County Selection
     level_list = ['State Level', 'County Level']
     level_selection = st.selectbox("**Geographic Level Selection**", level_list)
 
+    ## Display the Share of Renters Housing Burdened figs based on filter selection
     ast.renter_house_burden(level_selection, us_states)
 
-    # Section 3
-    st.write('')
-    st.markdown(f'#### National Average vs. Select State -- Year over Year')
-    st.write('')
+    
+    # Section 3: National Average vs. Select State -- Year over Year
+    ## Section title
+    tl.write_around_markdown('#### National Average vs. Select State -- Year over Year', 1, 1)
+
+    ## State Selection
     state_selection = st.selectbox("**Select State**", us_states, index=us_states.index('Colorado'))
 
+    ## Display the National Average vs. Select State -- Year over Year figs based on filter selection
     ast.yoy_comp_line_charts(state_selection)
 
     # Final Section
-    st.write('')
     ast.housing_terms()
 
 
+# Create Stock Market Page
 if view_selection == 'Stock Market':
     # Add n random s&p 500 stock tickers
     ss.stock_ticker(6)
 
     # Add ability to refresh page (gets new stock tickers)
-    st.sidebar.write('')
-    if st.sidebar.button('Refresh Page'):
+    if st.sidebar.button('Refresh Stock Tickers'):
         st.session_state.clear()
 
     # Add market index time series section
