+++ functions/stock_market/stocks_charts.py
